Map legacy translations of "base" to the actual base language
