+++ legacy/definition_test.go
@@ -269,31 +269,3 @@ func checkFlowLocalization(t *testing.T, flow flows.Flow, expectedLocalizationRa
 
 	assert.Equal(t, string(expectedLocalizationJSON), string(actualLocalizationJSON))
 }
-
-func TestTranslations(t *testing.T) {
-	// can unmarshall from a single string
-	translations := make(legacy.Translations)
-	json.Unmarshal([]byte(`"hello"`), &translations)
-	assert.Equal(t, legacy.Translations{"base": "hello"}, translations)
-
-	// or a map
-	translations = make(legacy.Translations)
-	json.Unmarshal([]byte(`{"eng": "hello", "fra": "bonjour"}`), &translations)
-	assert.Equal(t, legacy.Translations{"eng": "hello", "fra": "bonjour"}, translations)
-
-	// and back to JSON
-	data, err := json.Marshal(translations)
-	require.NoError(t, err)
-	assert.Equal(t, []byte(`{"eng":"hello","fra":"bonjour"}`), data)
-
-	translationSet := []legacy.Translations{
-		{"eng": "Yes", "fra": "Oui"},
-		{"eng": "No", "fra": "Non"},
-		{"eng": "Maybe"},
-		{"eng": "Never", "fra": "Jamas"},
-	}
-	assert.Equal(t, map[utils.Language][]string{
-		"eng": {"Yes", "No", "Maybe", "Never"},
-		"fra": {"Oui", "Non", "", "Jamas"},
-	}, legacy.TransformTranslations(translationSet))
-}
