+++ legacy/testdata/flows.json
@@ -26,7 +26,7 @@
                     "actions": [
                       {
                         "msg": {
-                          "base": "Hello",
+                          "eng": "Hello",
                           "fre": "Bonjour"
                         },
                         "media": {},
@@ -46,13 +46,6 @@
             "language": "eng",
             "expire_after_minutes": 0,
             "localization": {
-                "base": {
-                    "98388930-7a0f-4eb8-9a0a-09be2f006420": {
-                        "text": [
-                            "Hello"
-                        ]
-                    }
-                },
                 "fre": {
                     "98388930-7a0f-4eb8-9a0a-09be2f006420": {
                         "text": [
@@ -68,7 +61,7 @@
                         {
                             "type": "send_msg",
                             "uuid": "98388930-7a0f-4eb8-9a0a-09be2f006420",
-                            "text": "",
+                            "text": "Hello",
                             "attachments": []
                         }
                     ],
