+++ legacy/testdata/tests.json
@@ -59,6 +59,19 @@
     },
     {
         "legacy_test": {
+            "type": "contains_any",
+            "test": "oops not a dict"
+        },
+        "expected_case": {
+            "uuid": "d2f852ec-7b4e-457f-ae7f-f8b243c49ff5",
+            "type": "has_any_word",
+            "arguments": ["oops not a dict"],
+            "exit_uuid": "c072ecb5-0686-40ea-8ed3-898dc1349783"
+        },
+        "expected_localization": {}
+    },
+    {
+        "legacy_test": {
             "type": "contains_only_phrase",
             "test": {
                 "eng": "yes yeah",
